refactor(spa): extract SaveToastComponent to shared (T4f) [KAN-118]
Move save/dedup toast rendering from AppComponent into a self-contained
SaveToastComponent that subscribes to ToastService directly. AppComponent
now 109 TS + 41 HTML = 150 total lines (down from original 1,095+2,027).
