+++ src/app.component.html
@@ -37,48 +37,5 @@
 
   <app-footer />
 
-  <!-- Transient save/dedup confirmation toast -->
-  @if (saveToast(); as toast) {
-    <div
-      role="status"
-      aria-live="polite"
-      class="fixed bottom-4 inset-x-4 z-[60] flex justify-center md:inset-x-auto md:right-6 md:left-auto md:justify-end animate-[fadeIn_0.2s_ease-out]"
-    >
-      <div
-        class="flex items-center gap-3 rounded-2xl bg-stone-900 text-white shadow-2xl px-4 py-3 w-full max-w-sm"
-      >
-        <span class="flex-1 text-sm">{{ toast.message }}</span>
-        @if (toast.recipe) {
-          <button
-            type="button"
-            (click)="openSaveToastRecipe()"
-            class="text-sm font-semibold text-green-300 hover:text-green-200 underline underline-offset-2 whitespace-nowrap"
-          >
-            View
-          </button>
-        }
-        <button
-          type="button"
-          (click)="dismissSaveToast()"
-          aria-label="Dismiss notification"
-          class="text-stone-400 hover:text-white transition-colors"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            class="h-4 w-4"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              stroke-width="2"
-              d="M6 18L18 6M6 6l12 12"
-            />
-          </svg>
-        </button>
-      </div>
-    </div>
-  }
+  <app-save-toast (viewRecipeRequested)="viewRecipe($event)" />
 </div>
